XTT-955 Catch exception for unknown interface addr
We saw a marketplace crash where we couldn't get the network
interface address for a sit0 device, which is a device used to
get an IPv6 device working on IPv4 networks. I do not know a
setup nor device configuration to reproduce this, so this patch
will log the exception and carry on ignoring data from that device.